Bug 693463: Various small fixes.
Thanks to zeniko for these.

Use otf as extension for opentype fonts.

fz_clampi should take ints, not floats!

Fix typo in prototype.

Squash unwanted warning.

Remove magic number in favour of #define.

Reset generation numbers when renumbering.

+++ fitz/fitz.h
@@ -166,7 +166,7 @@ static inline float fz_clamp(float f, float min, float max)
 	return (f > min ? (f < max ? f : max) : min);
 }
 
-static inline int fz_clampi(int i, float min, float max)
+static inline int fz_clampi(int i, int min, int max)
 {
 	return (i > min ? (i < max ? i : max) : min);
 }
